+++ plugins/sudoers/log_client.c
@@ -1632,7 +1632,7 @@ handle_server_message(uint8_t *buf, size_t len,
  * XXX - share with logsrvd/sendlog
  */
 static bool
-expand_buf(struct connection_buffer *buf, unsigned int needed)
+expand_buf(struct connection_buffer *buf, size_t needed)
 {
     void *newdata;
     debug_decl(expand_buf, SUDOERS_DEBUG_UTIL);
@@ -1856,7 +1856,7 @@ client_msg_cb(int fd, int what, void *v)
     }
 
     sudo_debug_printf(SUDO_DEBUG_INFO,
-    	"%s: sending %u bytes to server", __func__, buf->len - buf->off);
+    	"%s: sending %zu bytes to server", __func__, buf->len - buf->off);
 
 #if defined(HAVE_OPENSSL)
     if (closure->ssl != NULL) {
@@ -1910,7 +1910,7 @@ client_msg_cb(int fd, int what, void *v)
     if (buf->off == buf->len) {
 	/* sent entire message, move buf to free list */
 	sudo_debug_printf(SUDO_DEBUG_INFO,
-	    "%s: finished sending %u bytes to server", __func__, buf->len);
+	    "%s: finished sending %zu bytes to server", __func__, buf->len);
 	buf->off = 0;
 	buf->len = 0;
 	TAILQ_REMOVE(&closure->write_bufs, buf, entries);
